display health

+++ Assets/Libraries/Bosco/IRandum.cs
@@ -17,7 +17,7 @@ namespace Bosco.Utils {
 	public interface IRandum {
 		bool NextBool();
 		double NextDouble();
-		long NextInt(long max);
+		int NextInt(int max);
 	}
 
 }
